implement ReviewManager API
