--- scopes/code/ui/code-compare-review/index.ts
@@ -1,10 +1,0 @@
-export { ReviewManager } from './review-manager';
-export type {
-  ReviewManagerSettings,
-  CodeSelection,
-  ReviewComment,
-  StyleSettings,
-  BaseStyleSettings,
-  DefaultIconRenderer,
-  OnChange,
-} from './models';
