+++ scopes/code/ui/code-compare/index.ts
@@ -1,4 +1,11 @@
 export { CodeCompare, CodeCompareProps } from './code-compare';
 export { CodeCompareTree, CodeCompareTreeProps } from './code-compare-tree/code-compare-tree';
-export { CodeCompareView, CodeCompareViewProps } from './code-compare-view/code-compare-view';
+export {
+  CodeCompareView,
+  CodeCompareViewProps,
+  CodeCompareReviewManagerProps,
+  CodeCompareViewState,
+  EditorType,
+  WithEditorType,
+} from './code-compare-view/code-compare-view';
 export { CodeCompareModel, CodeCompareContext, useCodeCompare } from './code-compare-context';
