+++ scopes/code/ui/code-review/context/review-manager-context/index.ts
@@ -1,0 +1,2 @@
+export { ReviewManagerProvider, ReviewManagerProviderProps } from './review-manager.provider';
+export { ReviewManagerContext, useReviewManager } from './review-manager.context';
