+++ scopes/code/ui/code-review/models/review-manager-model/index.ts
@@ -1,0 +1,16 @@
+export type {
+  IReviewManager,
+  OnChange,
+  ReviewManagerSettings,
+  ReviewComment,
+  CodeSelection,
+  StyleSettings,
+  BaseStyleSettings,
+  DefaultIconRenderer,
+  InternalReviewManagerSettings,
+  ReviewManagerEventType,
+  InitReviewManager,
+  InitReviewManagerProps,
+  ReviewEvent,
+} from './models';
+export { hasDefaultIconRenderer, styleDeclarationToString } from './models';
