+++ scopes/code/ui/code-review/review-manager/index.ts
@@ -1,0 +1,1 @@
+export { initReviewManager } from './review-manager';
