+++ scopes/component/code/code.ui.runtime.tsx
@@ -8,6 +8,8 @@ import { staticStorageUrl } from '@teambit/base-ui.constants.storage';
 import { CodePage } from '@teambit/code.ui.code-tab-page';
 import { ComponentCompareUI, ComponentCompareAspect } from '@teambit/component-compare';
 import { CodeCompareSection } from '@teambit/code.ui.code-compare-section';
+import { initReviewManager } from '@teambit/code.ui.code-review.review-manager';
+import { ReviewManagerProvider } from '@teambit/code.ui.code-review.context.review-manager-context';
 import { CodeAspect } from './code.aspect';
 import { CodeSection } from './code.section';
 
@@ -33,7 +35,33 @@ export class CodeUI {
   };
 
   getCodeCompare = () => {
-    return <CodeCompare fileIconSlot={this.fileIconSlot} />;
+    return (
+      // <ReviewManagerProvider
+      //   props={{
+      //     init: initReviewManager,
+      //     props: (props) => {
+      //       // console.log('🚀 ~ file: code.ui.runtime.tsx:43 ~ CodeUI ~ props', props);
+      //       return {
+      //         comments: [
+      //           { lineNumber: 1 },
+      //           { lineNumber: 6 },
+      //           { lineNumber: 3, selection: { startColumn: 6, startLineNumber: 3, endColumn: 6, endLineNumber: 5 } },
+      //         ],
+      //       };
+      //     },
+      //     onDestroy: (props) => {
+      //       // console.log('🚀 ~ file: code.ui.runtime.tsx:50 ~ CodeUI ~ props', props);
+      //     },
+      //     onChange: (props) => {
+      //       return (event) => {
+      //         // console.log(props, event);
+      //       };
+      //     },
+      //   }}
+      // >
+      <CodeCompare fileIconSlot={this.fileIconSlot} />
+      // </ReviewManagerProvider>
+    );
   };
 
   registerEnvFileIcon(icons: FileIconMatch[]) {
